requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==25.3.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.8
contourpy==1.3.1
cryptography==46.0.6
cycler==0.12.1
dash==3.0.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash_renderer==1.9.1
deprecation==2.1.0
distlib==0.4.0
fastapi==0.135.3
filelock==3.19.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
mysql-connector==2.2.9
mysql-connector-python==9.3.0
narwhals==1.31.0
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.0.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.2
pillow==11.1.0
platformdirs==4.4.0
plotly==6.0.1
postgrest==2.28.3
propcache==0.4.1
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pygame==2.6.0
Pygments==2.20.0
pyiceberg==0.11.1
PyJWT==2.12.1
PyMySQL==1.1.1
pyparsing==3.2.1
pypiwin32==223
pyroaring==1.0.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2024.1
pywin32==308
qrcode==8.2
realtime==2.28.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.3
retrying==1.3.4
rich==14.3.3
rpds-py==0.23.1
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
selenium==4.33.0
setuptools==78.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.41
starlette==1.0.0
storage3==2.28.3
streamlit==1.43.2
StrEnum==0.4.15
strictyaml==1.7.3
style==1.1.0
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.0.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
update==0.0.1
urllib3==2.4.0
uvicorn==0.43.0
virtualenv==20.34.0
watchdog==6.0.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
wordcloud==1.9.5
wsproto==1.2.0
yarl==1.23.0
zipp==3.21.0
zstandard==0.25.0
